[PAXWEB-1083] Clean up some issues
