+++ pax-web-itest/pax-web-itest-common/src/main/java/org/ops4j/pax/web/itest/common/AbstractSimultaneousWhiteboardIntegrationTest.java
@@ -48,6 +48,9 @@ public abstract class AbstractSimultaneousWhiteboardIntegrationTest extends ITes
 
 		simultaneousTestBundle.start();
 		whiteBoardBundle.start();
+
+		// It make some time for the services to get registered
+		Thread.sleep(1000);
 	}
