+++ pax-web-itest/pax-web-itest-common/src/main/resources/static/readme.txt
@@ -1,0 +1,1 @@
+registerResources test (static)
