+++ pax-web-itest/pax-web-itest-container/pax-web-itest-container-tomcat/src/test/java/org/ops4j/pax/web/itest/tomcat/HttpServiceIntegrationTest.java
@@ -63,7 +63,7 @@ public class HttpServiceIntegrationTest extends AbstractHttpServiceIntegrationTe
 	}
 	
 	@Test
-	@Ignore("worked before - framework settings?")
+	//@Ignore("worked before - framework settings?")
 	public void testNCSALogger() throws Exception {
 		testSubPath();
